episode 20

+++ app/Livewire/ArticleList.php
@@ -3,10 +3,11 @@
 namespace App\Livewire;
 
 use App\Models\Article;
-use Illuminate\Contracts\View\View;
 use Livewire\Component;
-use Livewire\Attributes\Title;
 use Livewire\WithPagination;
+use Livewire\Attributes\Title;
+use Livewire\Attributes\Computed;
+use Illuminate\Contracts\View\View;
 
 #[Title('Manage Articles')]
 class ArticleList extends AdminComponent
@@ -14,9 +15,26 @@ class ArticleList extends AdminComponent
     use withPagination;
     public $showOnlyTrashed = false;
 
+    #[Computed/*(persist:true)*/]
+    public function articles()
+    {
+        $query = Article::query();
+
+        if ($this->showOnlyTrashed) {
+            $query->where('published',1);
+        }
+        return  $query->paginate(10,pageName:'articles-page');
+        
+    }
     public function delete(Article $article): void
     {
+
+        if($this->articles->count()<10){
+            throw new \Exception('You must have at least 10 articles');
+        }
         $article->delete();
+        unset($this->articles);
+        cache()->forget(key:'published-count');
     }
 
 
@@ -31,17 +49,17 @@ class ArticleList extends AdminComponent
         $this->showOnlyTrashed = true;
         $this->resetPage(pageName:'articles-page');
     }
-    public function render(): View
-    {
+    // public function render(): View
+    // {
 
-        $query = Article::query();
+    //     $query = Article::query();
 
-        if ($this->showOnlyTrashed) {
-            $query->where ('published',1);
-        }
-        return view('livewire.article-list', [
-            'articles' => $query->paginate(10,pageName:'articles-page'),
-        ]);
-    }
+    //     if ($this->showOnlyTrashed) {
+    //         $query->where ('published',1);
+    //     }
+    //     return view('livewire.article-list', [
+    //         'articles' => $query->paginate(10,pageName:'articles-page'),
+    //     ]);
+    // }
 }
